LayoutTests/http/tests/security/xssAuditor/full-block-base-href-expected.txt

CONSOLE MESSAGE: line 1: Refused to load from document base URL. URL found within request.

There should be no content in the iframe below:



--------
Frame: 'frame'
--------
